Ensure DaemonSet pods are restarted when config changes
Since we're not using the Helm chart mechanism for configuring
fluent-bit, but rather provide our own configmap, the built-in
checksumming of the config in the Helm chart does not trigger pod
restarts when our custom config changes.

This commit introduces a second annotation on the DaemonSet podspec,
which is set to the MD5 sum of the contents of our custom config file
(syn-fluent-bit.conf in configmap syn-fluentbit-config).

This extra annotation should trigger pod restarts when our custom
configuration changes.