Preserve order of parser config files
Apparently, the order of parser config files matters sometimes, if one
parser depends on another parser which is defined in a different file.

This commit adjusts the logic which ensures that parser files are not
duplicated by replacing `std.set()` (which uses `std.sort()` internally)
with `std.foldl()`, to ensure the parser file order is preserved by the
component.

Custom parsers defined through `config.parsers` are added last in the
config.